+++ framework/source/nndeploy/op/expr.cc
@@ -127,7 +127,7 @@ std::shared_ptr<Expr> makeConv(ir::ModelDesc *model_desc,
   std::string name = op_name;
   if (name.empty()) {
     if (model_desc != nullptr) {
-      int index = model_desc->op_descs_.size();
+      int index = static_cast<int>(model_desc->op_descs_.size());
       name = "conv" + std::to_string(index);
     } else {
       name = "conv";
@@ -165,7 +165,7 @@ std::shared_ptr<Expr> makeSoftmax(ir::ModelDesc *model_desc,
   std::string name = op_name;
   if (name.empty()) {
     if (model_desc != nullptr) {
-      int index = model_desc->op_descs_.size();
+      int index = static_cast<int>(model_desc->op_descs_.size());
       name = "softmax" + std::to_string(index);
     } else {
       name = "softmax";
@@ -196,7 +196,7 @@ std::shared_ptr<Expr> makeAdd(ir::ModelDesc *model_desc,
   std::string name = op_name;
   if (name.empty()) {
     if (model_desc != nullptr) {
-      int index = model_desc->op_descs_.size();
+      int index = static_cast<int>(model_desc->op_descs_.size());
       name = "add" + std::to_string(index);
     } else {
       name = "add";
@@ -228,7 +228,7 @@ std::shared_ptr<Expr> makeRelu(ir::ModelDesc *model_desc,
   std::string name = op_name;
   if (name.empty()) {
     if (model_desc != nullptr) {
-      int index = model_desc->op_descs_.size();
+      int index = static_cast<int>(model_desc->op_descs_.size());
       name = "relu" + std::to_string(index);
     } else {
       name = "relu";
@@ -260,7 +260,7 @@ std::shared_ptr<Expr> makeSigmoid(ir::ModelDesc *model_desc,
   std::string name = op_name;
   if (name.empty()) {
     if (model_desc != nullptr) {
-      int index = model_desc->op_descs_.size();
+      int index = static_cast<int>(model_desc->op_descs_.size());
       name = "sigmoid" + std::to_string(index);
     } else {
       name = "sigmoid";
@@ -293,7 +293,7 @@ std::shared_ptr<Expr> makeBatchNorm(
   std::string name = op_name;
   if (name.empty()) {
     if (model_desc != nullptr) {
-      int index = model_desc->op_descs_.size();
+      int index = static_cast<int>(model_desc->op_descs_.size());
       name = "batchnorm" + std::to_string(index);
     } else {
       name = "batchnorm";
@@ -336,7 +336,7 @@ NNDEPLOY_CC_API std::shared_ptr<Expr> makeEmbedding(
   std::string name = op_name;
   if (name.empty()) {
     if (model_desc != nullptr) {
-      int index = model_desc->op_descs_.size();
+      int index = static_cast<int>(model_desc->op_descs_.size());
       name = "embedding" + std::to_string(index);
     } else {
       name = "embedding";
@@ -367,7 +367,7 @@ NNDEPLOY_CC_API std::shared_ptr<Expr> makeGemm(
   std::string name = op_name;
   if (name.empty()) {
     if (model_desc != nullptr) {
-      int index = model_desc->op_descs_.size();
+      int index = static_cast<int>(model_desc->op_descs_.size());
       name = "gemm" + std::to_string(index);
     } else {
       name = "gemm";
@@ -403,7 +403,7 @@ NNDEPLOY_CC_API std::shared_ptr<Expr> makeFlatten(
   std::string name = op_name;
   if (name.empty()) {
     if (model_desc != nullptr) {
-      int index = model_desc->op_descs_.size();
+      int index = static_cast<int>(model_desc->op_descs_.size());
       name = "flatten" + std::to_string(index);
     } else {
       name = "flatten";
@@ -433,7 +433,7 @@ NNDEPLOY_CC_API std::shared_ptr<Expr> makeMaxPool(
   std::string name = op_name;
   if (name.empty()) {
     if (model_desc != nullptr) {
-      int index = model_desc->op_descs_.size();
+      int index = static_cast<int>(model_desc->op_descs_.size());
       name = "maxpool" + std::to_string(index);
     } else {
       name = "maxpool";
@@ -462,7 +462,7 @@ NNDEPLOY_CC_API std::shared_ptr<Expr> makeGlobalAveragePool(
   std::string name = op_name;
   if (name.empty()) {
     if (model_desc != nullptr) {
-      int index = model_desc->op_descs_.size();
+      int index = static_cast<int>(model_desc->op_descs_.size());
       name = "globalaveragepool" + std::to_string(index);
     } else {
       name = "globalaveragepool";
